+++ ErsatzTV.Scanner/Core/Plex/PlexMovieLibraryScanner.cs
@@ -350,10 +350,11 @@ public class PlexMovieLibraryScanner :
             result.IsUpdated = true;
         }
 
-        if (fullMetadata.SortTitle != existingMetadata.SortTitle)
+        if (fullMetadata.SortTitle != existingMetadata.SortTitle || fullMetadata.Title != existingMetadata.Title)
         {
+            existingMetadata.Title = fullMetadata.Title;
             existingMetadata.SortTitle = fullMetadata.SortTitle;
-            if (await _movieRepository.UpdateSortTitle(existingMetadata))
+            if (await _movieRepository.UpdateTitles(existingMetadata, fullMetadata.Title, fullMetadata.SortTitle))
             {
                 result.IsUpdated = true;
             }