+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Update Plex movie and other video titles when changed from Plex
+  - The first library scan after updating will act like a deep scan due to adding title to the Plex etag calculation
+  - Future periodic scans will update titles in ETV automatically (deep scans will not be required to update titles)
 
 ## [26.6.0] - 2026-07-09
 ### Added